Replace order detail page with solvers view
